requirements.txt
pillow
drf-yasg
djangorestframework
django-filter